requirements.txt
aniso8601==9.0.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
gunicorn==20.1.0
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.1
opencv-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
